+++ Server.Test/ServerTest.cs
@@ -1,5 +1,6 @@
 using System;
 using Xunit;
+using Server.Model;
 
 namespace Server.Test
 {
@@ -11,8 +12,31 @@ namespace Server.Test
         }
 
         [Fact]
-        public void TestCRUD() {
-            
+        public void TestBooksCRUDForDb() {
+            // using (var db = new Context()) {
+            //         // Create and save a new Blog
+            //         Console.Write("Enter a name for a new Blog: ");
+            //         var name = Console.ReadLine();
+
+            //         var blog = new Blog { Name = name };
+            //         db.Blogs.Add(blog);
+            //         db.SaveChanges();
+
+            //         // Display all Blogs from the database
+            //         var query = from b in db.Blogs
+            //                     orderby b.Name
+            //                     select b;
+
+            //         Console.WriteLine("All blogs in the database:");
+            //         foreach (var item in query)
+            //         {
+            //             Console.WriteLine(item.Name);
+            //         }
+
+            //         Console.WriteLine("Press any key to exit...");
+            //         Console.ReadKey();
+            //     }
+            // }
         }
     }
 }