+++ Server/Misc/Config.cs
@@ -1,0 +1,8 @@
+
+namespace Server.Misc {
+
+    
+    public class Config : Common.Config {
+
+    }
+}