+++ Server/Model/Book.cs
@@ -1,0 +1,6 @@
+namespace Server.Model {
+
+    public class Book {
+        public string Title { get; set; }
+    }
+}