+++ Server/Model/Context.cs
@@ -1,0 +1,7 @@
+using System.Data.Entity;
+
+namespace Server.Model {
+    public class Context : DbContext {
+        public DbSet<Book> Books { get; set; }
+    }
+}